SubjectDirectoryAttributes, from Attributes to [+Attributes]
Following suggestion in issue 102:
"The ASN.1 is effectively list of tuples of OID and set. The CBOR is tuple of OID and set. There is no obvious way to map between the two. Furthermore, the present construction produces two values, which is not legal for extension, which can only produce one.
==> Wrap Attributes as [+Attributes]