SpecialText in attribute values - Take 2
The current optimization seems a bit strange as it bundles to optimizations together. This suggestion is a simplification.

This is a resubmission of #243, which was based on a now old version. Analysing #243, we found some inconsistencies that have been fixed.

A review comment from Joel on #243 was "I think this is an improvement!"